restructure imem
