+++ kernels/test_asm/dsip_bitstream.h
@@ -2065,7 +2065,7 @@ uint32_t dsip_rcs_imem_bitstream[4*DSIP_IMEM_SIZE] = {
   0x0,
   0x0,
   0x0,
-  0x0
+  0x0,
   0x0,
   0x420,
   0xc420,
@@ -2577,7 +2577,7 @@ uint32_t dsip_rcs_imem_bitstream[4*DSIP_IMEM_SIZE] = {
   0x0,
   0x0,
   0x0,
-  0x0
+  0x0,
   0x0,
   0x420,
   0xc420,
@@ -3089,7 +3089,7 @@ uint32_t dsip_rcs_imem_bitstream[4*DSIP_IMEM_SIZE] = {
   0x0,
   0x0,
   0x0,
-  0x0
+  0x0,
   0x0,
   0x420,
   0xc420,
@@ -3154,454 +3154,454 @@ uint32_t dsip_rcs_imem_bitstream[4*DSIP_IMEM_SIZE] = {
   0x0,
   0x0,
   0x0,
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
-  0x0
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
+  0x0,
 };
