+++ src/__init__.py
@@ -1,6 +1,6 @@
 from .lcu import *
 from .lsu import *
-from .ker_conf import *
+from .kmem import *
 from .mxcu import *
 from .rc import *
 from .imem import *
