+++ src/params.py
@@ -1,0 +1,4 @@
+# CGRA top-level parameters
+CGRA_COLS      = 2
+CGRA_ROWS      = 4
+N_VWR_PER_COL = 3
