+++ src/srf.py
@@ -1,0 +1,3 @@
+
+# Scalar register file
+SRF_N_REGS = 8
